requirements_self.txt
absl-py==2.1.0
addict==2.4.0
autocommand==2.2.2
backports.tarfile==1.2.0
basicsr==1.4.2
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
# Editable Git install with no remote (face_restoration==0.1.0)
-e e:\project\face_restoration\face_restoration
facexlib==0.3.0
filterpy==1.4.5
fonttools==4.54.1
future==1.0.0
gfpgan==1.3.8
grpcio==1.67.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
more-itertools==10.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.25.0
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.28.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
realesrgan==0.3.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
tb-nightly==2.19.0a20241019
tensorboard-data-server==0.7.2
tifffile==2024.8.30
tomli==2.0.2
torch @ file:///F:/software/torch-1.13.0%2Bcu117-cp39-cp39-win_amd64.whl#sha256=609017e9c9f0e20e6e4b82caab3c3c998e347bb179618777ef3570ca206bc283
torchvision @ file:///F:/software/torchvision-0.14.0%2Bcu117-cp39-cp39-win_amd64.whl#sha256=9daef60a1484389c261cac3e813232dccca7ed9491e08152471e7718131becb3
tqdm==4.66.5
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yapf==0.40.2
zipp==3.20.2
